frontend: Move (most of) ocm.go to the ocm package
This puts all the Cluster Service interaction code in one place.
